10-05 done

+++ src/n1008.java
@@ -1,14 +1,15 @@
+
 import java.util.Scanner;
 
 public class n1008 {
   public static void main(String[] args) {
     Scanner sc = new Scanner(System.in);
 
-    int A = sc.nextInt();
-    int B = sc.nextInt();
+    double A = sc.nextDouble();
+    double B = sc.nextDouble();
    
     System.out.println(A/B);
     sc.close();
-    
+
   }
 }
